pyproject.toml
[tool.poetry]
name = "application-for-analyzing-bank-transactions"
version = "0.1.0"
description = "A project for analyzing bank transactions"
authors = ["Evgenia <evgeniabalackaa6@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.4.1"
pytest-cov = "^6.0.0"
types-requests = "^2.32.0.20250306"
types-flask = "^1.1.6"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
)/
'''

[tool.isort]
line_length = 119
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
force_grid_wrap = 0
combine_as_imports = true
lines_between_types = 1
lines_after_imports = 2

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
